requirements.txt.txt
Django==2.2.1
django-heroku==0.3.1
djangorestframework==3.9.4
gunicorn==19.9.0
psycopg2==2.8.2
uWSGI==2.0.18
whitenoise==4.1.2%  